+++ extensions/envato/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Envato Changelog
-
+## [Verify Purchase] - 2023-10-11
+- Added Verify Purchase using Purchase Code
 ## [Update] - 2023-05-02
 
 - Tracking New Sales For Envato Author